+++ packages/react-tinacms-strapi/src/index.ts
@@ -10,10 +10,15 @@ You may obtain a copy of the License at
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
 
 */
 
-export * from './Dismissible'
+export * from './AsyncButton'
+export * from './popupWindow'
+export * from './strapi-auth-modal'
+export * from './strapi-media-store'
+export * from './strapi-provider'
+export * from './strapi-client'