+++ packages/gatsby-tinacms-remark/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.24.1-alpha.0](https://github.com/tinacms/tinacms/compare/v0.24.0...v0.24.1-alpha.0) (2020-07-22)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
 ## [0.8.7-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.8.6...gatsby-tinacms-remark@0.8.7-alpha.0) (2020-07-15)
 
 **Note:** Version bump only for package gatsby-tinacms-remark