+++ packages/react-tinacms-editor/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.24.1-alpha.0](https://github.com/tinacms/tinacms/compare/v0.24.0...v0.24.1-alpha.0) (2020-07-22)
+
+
+### Bug Fixes
+
+* UX improvements hide title input from link modal ([6e5ab20](https://github.com/tinacms/tinacms/commit/6e5ab20))
+
+
+
+
+
 ## [0.8.5-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-editor@0.8.4...react-tinacms-editor@0.8.5-alpha.0) (2020-07-15)
 
 **Note:** Version bump only for package react-tinacms-editor